# Cache host function static type