@turbolent 2021-10-05
@SupunS Would you be able to post a benchmark comparison, just to make sure this does not actually regress performance? I'm worried the locking might have a big overhead, potentially more than the gain from caching