@SupunS 2021-10-05
Yeah, checking that now. Also removing locking means we have to remove the base-activation which is a singleton now, which might affect runtime performance (we made it a singleton as an optimization). Running a benchmark for that as well