@SupunS 2021-10-06
@turbolent These are the benchmark results with locks. (_old_ - with locks, _new_ - without locks)
```
name                                  old time/op    new time/op    delta
FunctionStaticType/Get_type-8           2.12ns ± 1%    2.10ns ± 1%     ~     (p=0.063 n=4+5)
FunctionStaticType/Get_inner_types-8    5.75ns ± 0%    3.71ns ± 1%  -35.58%  (p=0.016 n=4+5)

name                                  old alloc/op   new alloc/op   delta
FunctionStaticType/Get_type-8            0.00B          0.00B          ~     (all equal)
FunctionStaticType/Get_inner_types-8     0.00B          0.00B          ~     (all equal)

name                                  old allocs/op  new allocs/op  delta
FunctionStaticType/Get_type-8             0.00           0.00          ~     (all equal)
FunctionStaticType/Get_inner_types-8      0.00           0.00          ~     (all equal)
```

But that would mean we have to make the base activation non-singleton. This means newing the interpreter takes a lot of time:
```
name                                  old time/op    new time/op    delta
NewInterpreter/new_interpreter-8        1.18µs ± 3%   15.37µs ± 1%  +1197.32%  (p=0.008 n=5+5)
NewInterpreter/new_sub-interpreter-8    1.86µs ± 1%   16.32µs ± 3%   +779.10%  (p=0.008 n=5+5)

name                                  old alloc/op   new alloc/op   delta
NewInterpreter/new_interpreter-8          912B ± 0%    16711B ± 0%  +1732.39%  (p=0.008 n=5+5)
NewInterpreter/new_sub-interpreter-8    1.25kB ± 0%   17.05kB ± 0%  +1265.95%  (p=0.008 n=5+5)

name                                  old allocs/op  new allocs/op  delta
NewInterpreter/new_interpreter-8          14.0 ± 0%     286.0 ± 0%  +1942.86%  (p=0.008 n=5+5)
NewInterpreter/new_sub-interpreter-8      33.0 ± 0%     305.0 ± 0%   +824.24%  (p=0.008 n=5+5)
```
Basically undoing the changes in: https://github.com/onflow/cadence/pull/1018