@turbolent 2021-11-08
@SupunS and I synced offline:

The first benchmark compares:
- current master: always re-computing, no caching, and no locks
- this PR: caching, with locks
The later is slower. 

The second benchmark is for an alternative to caching, where we don't use locks, but then have to undo the base activation optimization. This would be a dramatic regression

We decided to close this PR and leave the current non-caching code on master as-is.